requirements.txt
easygui==0.98.3
Pillow==9.1.1
